Flaw in quality formula ultimately causes gaining expertise to cause a drop in product quality to wrap to 0.
In quite common circumstances, gaining expertise caused the total sum of
ability * expertise contribution to exceed 10. The mod 10 would then
cause a wrap around from 10 to 0.
In the fix, the average expertise contribution (between 0 and 1) is used
to scale the mod 10 sum of abilities. As such, the quality will increase
upon gaining expertise and always be >=0 and < 10.